created the seperate file for making the http calls
