+++ DotNet Examples/TestMonitor/DeleteResults/TestDataManagerDeleteExtensions.cs
@@ -37,7 +37,7 @@ namespace NationalInstruments.SystemLink.Clients.Examples.TestMonitor.DeleteResu
             }
             string path = $"{DeleteResultPath}/{resultId}?deleteSteps={deleteSteps}";
 
-            return CallDeleteResultApi(configuration, path);
+            return TestMonitorHttpClient.RaiseDeleteRequest(configuration, path);
         }
 
         /// <summary>
@@ -60,75 +60,18 @@ namespace NationalInstruments.SystemLink.Clients.Examples.TestMonitor.DeleteResu
             {
                 Ids = resultIds.ToArray(),
                 DeleteSteps = deleteSteps
-            };
+            };            
 
-            var body = GeneratePostRequestBody(bodyContent);
-
-            return CallDeleteResultsApi(configuration, body);
+            return TestMonitorHttpClient.RaisePostRequest(configuration, bodyContent, DeleteResultsPath);
         }
 
-        /// <summary>
-        /// calls the delete result API
-        /// </summary>
-        /// <param name="configuration">A HttpConfiguration object containing base url and api_key</param>
-        /// <param name="requestUri">Uri route</param>
-        /// <returns>Http reponse message after calling the API</returns>
-        private static HttpResponseMessage CallDeleteResultApi(IHttpConfiguration configuration, string requestUri)
-        {
-            HttpClient client = GetHttpClient(configuration);
+        
 
-            var response = client.DeleteAsync(requestUri).Result;
-            CheckForSuccessResponse(response);
 
-            return response;
-        }
+        
 
-        private static void CheckForSuccessResponse(HttpResponseMessage response)
-        {
-            if (!response.IsSuccessStatusCode)
-            {
-                throw new HttpRequestException("Error occured while deleting the result/results. Please check for the correct result Id/Ids");
-            }
-        }
+        
 
-        /// <summary>
-        /// generates the body for the post request
-        /// </summary>
-        /// <param name="content">an object representing what to there in the body</param>
-        /// <returns>A string content which can be used as a body for the post request</returns>
-        private static StringContent GeneratePostRequestBody(object content)
-        {
-            return new StringContent(JsonConvert.SerializeObject(content), Encoding.UTF8, "application/json");
-        }
-
-        /// <summary>
-        /// calls the delete results APi
-        /// </summary>
-        /// <param name="configuration">A HttpConfiguration object containing base url and api_key</param>
-        /// <param name="body">request body</param>
-        /// <returns>Http response message after calling the API</returns>
-        private static HttpResponseMessage CallDeleteResultsApi(IHttpConfiguration configuration, HttpContent body)
-        {
-            HttpClient client = GetHttpClient(configuration);
-
-            var response = client.PostAsync(DeleteResultsPath, body).Result;
-            CheckForSuccessResponse(response);
-
-            return response;
-        }
-
-        /// <summary>
-        /// Generates the Http client for making the request
-        /// </summary>
-        /// <param name="configuration">A HttpConfiguration object containing base url and api_key</param>
-        /// <returns>HttpClient</returns>
-        private static HttpClient GetHttpClient(IHttpConfiguration configuration)
-        {
-            var client = new HttpClient();
-            client.BaseAddress = new Uri(configuration.ServerUri.AbsoluteUri);
-            client.DefaultRequestHeaders.Add("x-ni-api-key", configuration.ApiKeys["x-ni-api-key"]);
-
-            return client;
-        }
+        
     }
 }
